Add README
Describes some features that are not yet available but will be in the future. A good place to brain dump the overall idea.

Main thing to do is to create commit and create a PR when there are changes. Then it can be used for SwiftPM projects.

Xcode projects later since the project should be validated as part of the PR and that might be tricky.